+++ src/route-linux.c
@@ -172,7 +172,7 @@ route_get(route_t *r, struct route_entry *entry)
 	if ((i = recvmsg(r->nlfd, &msg, 0)) <= 0)
 		return (-1);
 
-	if (nmsg->nlmsg_len < sizeof(*nmsg) || nmsg->nlmsg_len > i ||
+	if (nmsg->nlmsg_len < (int)sizeof(*nmsg) || nmsg->nlmsg_len > i ||
 	    nmsg->nlmsg_seq != seq) {
 		errno = EINVAL;
 		return (-1);
